requirements.txt
Flask==3.1.2
pymongo==4.15.4
gunicorn==23.0.0

numpy==2.0.1
scipy==1.15.3
opencv-python==4.12.0.88
pillow==10.4.0
matplotlib==3.10.7

ultralytics==8.3.225
ultralytics-thop==2.0.18
